test(optid): answer the enabled-versus-off question with simulated evidence
The existing I2 simulation compares a scenario manifest against expected
results the same manifest declares. That proves the matrix engine, not optid.
This adds a harness that drives the unmodified production control loop —
`crate::run` — against a machine materialised inside a verified simulation
root, and computes the modelled outcome from the machine state optid actually
left behind.

Real optid code performs the sensing, workload classification, mode selection,
per-domain gating, hardware-allowlist and contract and capability checks,
transactional actuation, journalling, circuit-breaker accounting, shutdown
restoration and startup recovery. The real S3D `optid-recover` executable runs
as a subprocess before every supervised restart, matching
`optid-apply.service`'s `Requires=optid-recover.service`. Only the machine
underneath is modelled, by a model that has no knowledge of which arm is
running.

Entry point: `optid --evidence-root <dir>`, present only in a build with the
non-default `test-simulation` feature. A normal build has no such surface.

Containment: every path is either already inside the verified root or is
rewritten into `<run>/machine/<path>` before any syscall; anything else is
refused with EPERM and recorded. `PATH` is emptied, the D-Bus addresses point
at a dead socket, and the competing-daemon detector is pinned, so no host
binary and no system service can be reached. The run reports zero write
attempts outside the root.

Production seams extended so the simulated machine is reachable through the
existing F2 boundary rather than around it: the backlight, runtime-PM and
storage predicates and the allowlist override load and HWID resolution now
read through `KernelRead` instead of `std::fs`; the `RealKernel` override, the
conflict-checker override and a new PM QoS sink override compile under the
feature as well as under `cfg(test)`; `optid-recover` gains a
`--machine-root` flag that exists only in a feature build.

Evidence and report: release/evidence/optid-simulated-evidence/2026-08-29/

Also fixes a pre-existing `unnecessary_cast` clippy error in
crates/testos/src/run_intent.rs that blocked
`cargo clippy --workspace --all-targets -- -D warnings`.

Ledger: updates only I2, which stays `candidate`. It is not marked completed.

+++ crates/testos/src/run_intent.rs
@@ -427,7 +427,7 @@ fn parse_iso8601_to_epoch(s: &str) -> Option<u64> {
     let doy =
         (153 * (if month > 2 { month - 3 } else { month + 9 }) as u64 + 2) / 5 + (day - 1) as u64;
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy; // [0, 146096]
-    let days_since_epoch = era as i64 * 146097 + doe as i64 - 719468;
+    let days_since_epoch = era * 146097 + doe as i64 - 719468;
     if days_since_epoch < 0 {
         return None;
     }
